Remove unused byte
The message data in quick_use() contained an extra byte which is never used. The byte containing the position
of the item to use in the inventory, was added in 2005 but, as far as I can tell, was never sent to the server.
Since the server hasn't missed it in 16 years, and regular use from the inventory does not send such a byte either,
let's simply remove it. It can be reintroduced with proper explanation when the server can actually use it.