Dynamically allocate servers list
Currently there is a fixed upper limit on the number of servers in the servers list. Most people will probably
never reach this limit, but in testing different connections I have had to raise it twice now. So let's
allocate the servers list dynamically, so that people can have an arbitrary number of servers in their config.