Add (optiona;) encryption field to servers.lst
Add an optional Encryption column to the servers.lst file between the Port and Description fields that
determines whether the connection should be encrypted. If the columns contains "crypt", "encrypt", or "encrypted",
the connection should be encrypted, if it contains "plain" or "clear", or is left empty, the connection is
not encrypted. The match is case-insensitive. Please note that the field is optional, so any other value
than the 5 listed above will be parsed as part of the description. Conversely, server descriptions should
not start with any of the five reserved words unless the encryption column is explicitly set.

Since the field is optional, existing servers,lst files should still work. Note that the new field is only
parsed when the client is compiled with USE_SSL defined. Otherwise, any contents of the encryption
column will be added to the description.