Add C++ abstraction for windows
Add a C++ Window class that wraps the C window_info struct, and can automatically register event handlers
and handle capturing lambdas as callbacks. This class does not yet expose the full set of window and
widget capabilties, and will probably evolve further. But it is a base to start from.