Add text popup
Add a new class for text popups that can display a text message and a set of buttons with various callbacks
underneath. This is heavily inspired by the serverpopup window, but adds the ability to add multiple buttons.
The server popup could be trivially reimplemented in terms of this class.