Clean up connection before exit.
Welcome to the wonderful world of global statics. The destructor of the Connection singleton may be called
after the context menu container singleton. This is a problem since the Connection can contain a (popup) window,
that will try to clean up its (non-existant) context menu upon destruction, and tries to look itself up
in the destroyed container. Ensure the popup is destroyed before the program finsihes.