Generate password keys more securely.
When using SSL, generate the key for the password file using a cyptographically secure random number generator.
Otherwise, on non-Windows systems, try to seed the random number generator with a truly random seed. Doing
the same on Windows is a bit problematic: rand_s() is not supported by gcc on MSYS, using RtlGenRandom()
directly is a pain, and pulling in the full crypto system (either the old and deprecated one or the new one)
feel like overkill if we already have OpenSSL support. So for now, the situation on Windows remain unchanged
when USE_SSL is not enabled.

This only affects the generation of new password keys. Also the key format hasn't changed, so existing clients
can decode newly generated keys.