Enable packet compression
Learner's proxy (and the OL server) are able to large (or multiple small) messages to cut down on network usage
when much data needs to be sent at once (e.g. on login, map change, or when many things are happening at the
same time). The EL server has never supported this, but with encrypted connections now connecting to a proxy
that has support for it, we can enable it in the client too. The compression support is enabled when connecting
to the server, by sending an empty OL_COMPRESSED_PACKET message to the server. This is only done for encrypted
connections, so as not to send unknown opcodes the normal server.