Fix compilation on windows
Somehow, the mingw compiler in msys2 manages to thoroughly confuse itself with the use of X509_NAME type definition.
Compilation fails with

 error: 'name' was not declared in this scope; did you mean 'tzname'?

which is completely bogus. Interestingly when the type of `name` is changed to `void*`, it can deduce that in the
next line where it is used as an argument, it should have a type `X509_NAME*` (with the properly expanded
definition). I have no idea what's going on here, it seems to be unique to the gcc 10.1.0 compiler on msys2.
Neither gcc 11.1.0 on Linux, nor Clang on FreeBSD, seem to have an issue. Expanding the typedef seems to solve
the issue.