Ask server for encryption.
Learner prefers to support both encrypted and unencrypted connections on a single port. This means we cannot
blindly encrypt the whole session as we did previously, but somehow have to tell the server through an EL
packet that we want to encrypt the session. So we revert to the initial design: on startup, the client sends
an empty LETS_ENCRYPT packet to server. The server responds with its own LETS_ENCRYPT message, carrying a single
byte indicating whether encryption is supported. If the server response is negative, or when no response arrives
within 5 seconds, the connection is broken with an appropriate error message.

Unfortunately this is a protocol change, which means that all clients currently using the encryption functionality
must update, and that the new code needs to be tested again.