Clear "waiting for encryption response" flag on connect.
When the client requests encryption from the server, but disconnects before it can process the response, the
corresponding flag remains set. This has the effect that on reconnect, no new invitation can be sent because
the client thinks it is still waiting. This in turn leads to the server/proxy never sending its own LETS_ENCRYPT
message, with the result that the client disconnects again after 5 seconds, and the whole process repeats, so
that the client never makes a succesfull connection.

Clear the flag on reconnect, and only set it again after the encryption request is actually sent out to prevent
this endless loop.