Fix crash when window is minimized in new character interface
When the window is returned after it is minimized (or when changing to a virtual desktop on *nix, or really
anything that triggers SDL_WINDOW_HIDDEN or SDL_WINDOW_MINIMIZED to be set), the client tries to synchronize
all actors by destroying them all and requesting the server to send an updated list. This of course does not
work well with the new character interface, which expects its single actor to remain available (and which
cannot sync with the server anyway).

So, do not try to update the actors when the new character window is visible.