Remove unrestricted access to actors_list
Make locking madatory for access to the actors_list. This is done by introducing a new singleton class ActorsList
that holds the actual actors storage as a private member, and only provides access to this list by member functions
that lock a mutex.

As the actors list is deeply ingrained into the game, this patch has is large, and may introduce bugs. To help with
debugging, a new compile option ACTORS_LIST_MUTEX_DEBUG is introduced. This option turns the mutex into a timed
mutex, and calls did_we_deadlock() if the mutex cannot be locked within a second. Setting a breakpoint on
did_we_deadlock() can help reveal where the deadlock occurs.

This work is not fully done yet, as access to your_actor still needs to be protected.