Get rid of your_actor
Get rid of the global pointer to the player's own character, which in many places was accessed without taking
the actors list lock. Unfortunately, at least for the time being, this required switching to a recursive mutex,
as in add_sound_object_gain() the actor's current position is required. This function is deep in the call chain,
and can be called from different contextxs, so passing the position (or pointer to the actor) down becomes
unwieldy.