Update Pawn code to use new actors list
Make the Pawn code use the new actors list. In addition, the existing code returned the index into the actors list
when looking up an actor by name. This is a bad idea, as this index can change when actors are removed. Fix and
return the actor ID instead.

Not that this will matter one whit, as Pawn was never enabled in a release build, and support for it will probably
be removed in the not too distant future.