Replace actor pointers in ec_internal_reference with actor IDs
The eye candy wrapper kept pointers to actor structures in its own internal structures. These pointers could
be invalidated when the actor is removed, though this seems to be handled by the fact that ec_actor_delete()
is called when the actor data is freed. Nevertheless, let's keep the actor pointers stored in a single place
(i.e. the actors list), and replace the pointers by actor IDs. If anything, it makes the ec_internal_reference
struct a bit smaller.