Replace actor pointer in struct special effect with ID
Struct special_effect can store a pointer to the actor it affects, but this pointer can be invalidated when the
actor is removed from the actors list. This is taken care of by the fact that free_actor_special_effect() is
called when the actor is destroyed, but let's keep the actor pointers in a single place. Replace the pointer with
an ID, and retrieve the actor from the actors list when it is needed.