Don't rely on indexes in the actors list when determining actor under mouse cursor
When actors are removed from the actors list, the last actor in the list is moved into the empty spot. This means
positions in the actors list can change, and code should not rely on these indexes if the actors list lock
is released between determining the index and using it. Fix this for the code that determines what's under
the mouse cursor by using the actor ID instead of the index.