Fix initial state of fighting horses
If a new actor is sent from the server which is both fighting and has a horse, the horse needs to be marked as
fighting as well. To check if there is a horse present, the attached_actor field of the main actor was checked, but
that was at that point still unset since the actor hadn't been added to the actors list yet. Check instead if
an attached actor was created.