Stop relying on indexes in the actors list
The index of an actor in the actors list can change if the actors list lock is released between determining the
index and using it. Stop relying on indices, and store actor IDs to identify actors. To that end, horses now
also get a (fake) ID, starting at 0x10000. Since server IDs are two bytes only, this should not clash with
actor IDs from the server.