Stop relying on representation of actor's list
Currently, the code relies on the actrs list being stored as a continuous array of pointers to actors. This
is used in a few places for iterating over the actors list. Add a pair of function to iterate over the actors
list with a callback (two functions for with and without attached actor), This makes the actual storage of
the actors in the list an internal implementation detail.