Make Locked class lock the mutex again by default
There is no need to lock the mutex before calling the Locked constructor when protecting the actors list itself.
Make the constructor of the Locked class lock the mutex again, and add a second constructor with a std::adopt_lock_t
argument (like std::lock_guard has) which can be used when the mutex is already locked.