Allow 'm' as million suffix in #calc
There is already support for 'k' to indicate multiplication by 1,000, add 'm' as a number suffix to indicate
multiplication by 1,000,000. There is a slight disadvantage that 'm' is also used for the memory cell of the
last result. The syntax is not ambiguous, as this memory cannot occur after a number, but it may be surprising
to some to see

1 = 1
m = 1
m+m = 2
m+2m = 2000002

However, the familiarity of the 'm' suffix IMO outweighs this possible problem.