Rename proxy_ext.[ch] to ext_protocol_shared.[ch]
The commands handled by this file are not only sent by Learner's proxy servers, but also by the Other Life
servers themselves. Perhaps one day they will be incorporated into the EL server as well. Rename the file
to reflect this reality, and to facilitate the merge with the OL client code.