Fix byte count off by one in trade withdrawal
When withdrawing an item from trade, only the first three bytes of the amount are sent. The server does not
seem to care, as noone has ever reported withdrawing items from a trade does not work. Since EL data is little
endian, it should only matter when huge numbers are withdrawn, but let's do what is right, and send the
entire amount.