Clamp light color values in eye candy effects
The calculated color values for the lights in the eye candy effects may go above 1.0, which is invalid. Clamp
the values to a maximum value of 1.0 to fix this. The invalid color values may be the cause of the bug where the
screen goes dark in heavy combat scenes where a lot of magic is used. Reproducing this bug is a bit hit and miss,
it does not always occur, but with this patch, I have not been able to trigger it at all anmore.