Band-aid fix for invalid particle coordinates.
When updating particle positions for a target magic effect, an effect-dependent shift is added to the
particle's position. This shift is only valid for particles belonging to a TargetMagicEffect2 object.
But not all TargetMagicParticles do, and for these articles casting the pointer to their effect to a
TargetMagicEffect2 pointer results in an invalid memory access with (most likely) entirely invalid
values. Check if the effect is indeed a TargetMagicEffect2 before adding the shift to fix this.

This is probably not the most efficient solution to this problem. Fixing it properly would require some
more intense refactoring.