Fix crash on Adreno
On Android systems with an Adreno GPU, the game crashes when drawing both
the player name and damage updates in the player banner. This is most
commonly seen when fighting. I do not know the exact root cause of the
problem (whether it is in the client code, gl4es, or the Adreno drivers)
but checking the depth buffer also when drawing the damage updates seems
to prevent the crash from happening.