Safely fill filename buffer in object browser.
Ensure the filename buffer in the object browser does not overflow using safe_strncpy() and safe_strcat().
An attempt to do so was already done for the first (of four) objects in the window but the third argument
to strncat() means the maximum number of bytes to take from the source string, not the maximum size of the
destination buffer. In safe_strcat(), it does.