Set color to white before drawing objects in object browser.
In the object browser, the GL draw color was set to the UI color to draw the UI elements, and not reset
before drawing the 3D objects in the browser window. This had the effect of drawing the first object in
the window with a brownish tint, as the 3D object code enables GL_COLOR_MATERIAL to track the current
draw color as ambient and diffuse light colors, and enabling the GL_COLOR_MATERIAL bit will implicitly
set these colors to the current GL color. The other (up to) 3 objects in the browser window were not
affected, since the 3D object drawing code resets the current GL color to white after drawing the object,
and when these objects enable GL_COLOR_MATERIAL, the light colors are reset to white as well.

However, this reset was done *after* the GL_COLOR_MATERIAL was disabled again. So if there was only a
single object to draw in the browser window (e.g. at the end of the list), the light colors were not
reset after drawing the object. This had the effect of drawing the game map in the main window (though not
the 3D objects on it) in a brownish light. Setting the GL color to white before drawing any 3D objects
fixes this.

I am not sure whether resetting the GL color to white in the 3D object drawing code is the correct choice.
I have left it in place for now, but if it is done at all, it should probably be moved up to occur before
GL_COLOR_MATERIAL is disabled.