Limit "You cannot access the storage from here" messages
Issue #161: when clicking the "store all" button in the inventory, while being out of range of the storage,
the server will send an error message for every item in the storage, spamming the client. Show only one of
the per click on the button (or once per 5 seconds if the messages arrive *really* slowly).

This should really be handled by the server (as well as the other rate limited messages).