Fix water reflection on Windows/Virtualbox/Intel graphics
Fix a number of issues with the water reflection running the client on a Windows client in Virtualbox:
* The framebuffer texture is interpreted upside down by the VirtualBox driver. This is an issue that has been
encountered in other programs as well, notably on Intel integrated graphics chipsets. Fix this by detecting
if this is the case by generating a known texture and checking its contents, and flipping the scene vertically
if it occurs.
* The OpenGL implementation seems to have trouble with the water shader using a vectored multiply-add-assign
operation. Explicitly constructing a new vec2 from separate components seems to solve the issue somehow. Though I
am no expert in GLSL, the existing code certainly looks valid, so this seems to point to another driver bug.