Prepend date to chat log also when timestamps are shown in game.
When timestamps are shown in game, the text to be logged already contains a timestamp and no date was prepended.
Fix this by adding the current date to the log buffer and copying the rest of the timestamp from the logged message.
There is a tiny chance that when the day rolls over the timestamp is inconsistent with the date (i.e. we get the date
part of the next day with a late timestamp of the previos day). However, this should be very infrequent, and doing it
this way ensures the logged timestamps are consistent with what is shown in game.