Don't approximate destination if path becomes too long
In commit d95ff65e, the pathfinder was changed to move to a nearby location if an unwalkable tile is clicked. A side
effect of this change, is that it becomes easier to click on an unwalkable tile in a river, fence, hedge, or other
such object, and start a long unwanted walk around it as a result. For that reason, restrict the walk to an approximate
location to at most three times the length of a direct walk to the clicked (unwalkable) destination tile.

Pathfinding to a walkable tile remain unaffected, so that it remains possible to walk to the other side of a fence
without by simply clicking there.