Add explicit close() and error logging when closing JSON files. Several people have reported issues with the JSON files, specifically format errors when reading the file.  When the error handling was added, it was indented to spot if someone had edited the file and broken the structure.  I've been unable to get anyone to send an example file that would enabling debugging.  Someone did say that the corrupted file was zero byes in length.  This could indicate an error when writing the file. Previous the closing the file was left to the object destructor, now we explicitly close the file and check and log errors.
Possible cause for the error.  In each case, the JSON object and the
file are created on the function stack, the ordering ensuring that
the file is destructed before the JSON object destructor, that's the
standard as I understand it.  But if that is not the case, then the object
could be empty and a zero length file would be written.  Adding the
explicit close() will prevent this.