Improve destination highlighting
Currently, when clicking the world map while in walk mode, a highlight is displayed on the clicked ground tile
regardless of whether a walk to that tile can be initiated or not. Furthermore, with the pathfinder now able
to plot a path to a nearby location if the destination tile is not walkable, the highlight may not indicate
the position the actor is actually walking to.

The second problem is fixed by actually highlighting the destination tile found by the pathfinder. As for the
first problem: when the pathfinder fails to find a path, or when the pathfinder is bypassed but the clicked tile
is unwalkable, it is very unlikely that the server will initiate a walk to that destination, so don't display
a highlight in that case.