Fix noise testure for Core profile >= 3
The noise texture for the water ripples is stored in GL_LUMINANCE_ALPHA format. Unfortunately, OpenGL 3 deprecated
GL_LUMINANCE_ALPHA, and mandates the use of GL_RG instead, On the other hand, OpenGL <= 2 does not define GL_RG.
In order to support both, we first try to use GL_LUMINANCE_ALPHA, and if that fais with a GL_INVALID_ENUM error,
try GL_RG instead (and swizzle the texture so the shader still sees the correct values from the green and alpha
channels). If that fails too, water ripples are disabled.