Add new water shader
Add a new water shader which is compatible with GLSL 1.50. Some systems no longer support GLSL 1.40 and lower
(specifically many of gl_* variables), and so cannot render the water reflection and rippling. The new shader
should work on such systems. Keep the old shader around for ancient or crippled systems.