Save state when checking reflection orientation.
Commit 71a2f437 intorduced a test to see if the reflection texture was read upside down. This test did not save the
current OpenGL state, notably the depth test and lighting flags. This broke the new character scene, where due to the
lack of depth testing parts of the scene could disappear behind others where they should not. Fix this by saving the
state and restoring it at the end of the function.