Add eye candy lights.
Add the lighting due to eye candy effects to the new water shader. Unfortunately, this does not matter one whit at
the moment, as
1) in daytime, no eye candy lights are used, and
2) in nighttime, no shadows are drawn, and
3) the lighted parts are unaffected, because we don't use the actual light color but solid white like the old shader does.

Well, at least when when you adjust the new water shader to use the actual light color, the water will light up in
nighttime when you eye candy effects are performed near it. We may want to revisit this at some point.