Fix scene lighting
Commit b4432448 Stored the sun's world position in world coordinates for use with the new water shader. Unfortunately,
in doing so, the "w" component of the global light position was changed from 0.0 to 1.0. This meant it was nog longer
interpreted as a directional light but as a positional light to which light attenuation wast to be applied. Because
the constant attenuation factor of the global light was set also set zero, the light scaling factor blew up, and the
scene was far too bright in night time.

Fix by copying the w-component for the light position as well.