Check if deprecated function works
Some systems when using an OpenGL 3.0 or higher context, do not support pre-3.0 functionality, even though a
compatibility profile is selected (for OpenGL >= 3.2), the ARB_compatibility extension is supported (for OpenGL 3.1),
or OpenGL 3.0 is used which deprecated this functionality but should not remove it.

Do a quick test to see if a deprecated function (glGetString(GL_EXTENSIONS)) can be used, and if that does not
work, choose a lower version context.