Reset number of enabled local lights on map change.
The number of enabled local lights was not reset on a map change. Since in daytime, no local lights are drawn on
outside maps, this variable was also not reset by calling draw_lights() as it would for inside maps or in night time.
Therefore, when moving from inside a building to the outside could result in the wrong number of enabled local
lights (i.e. more than 0 in day time), which in turn could crash the client.

Fix by also resetting the number in disable_local_lights(), which is called whenever local light no longer are drawn
(including on a map change to an outside map in day time).