Move UVP debug string under DEBUG #ifdef
Issue #183. When the actor animation program is used, a text message UVP: 1 is printed under the FPS string on the
main screen. In normal gameplay it is distracting, and unnecessary. Remove it for releases, but keep it as a DEBUG
option in case anyone wants to work on modernizing the associated code.