Dynamically allocate the near actors list
The current list of actors to be drawn is staically allocated, with an upper cap of MAX_ACTORS (1000). Change this
to a dynamically allocated array. This removes the last theoretical limit on the number of actors in the client
(though in practice drawing over 1000 actors will be painful, and we are still limited to 16-bit actor IDs from
the server).