Send the server port in network byte order
As part of the version information, the client sends the IP address and port of the server it connects to. While the
EL network protocol is little-endian, the server expects the port number to be in network byte order. In fact, when
the old (pre-SSL) network code is used, this is done, though it is somewhat obscured by the fact that SDL_net's
IPadress struct also stores the port number in network order, which looks reversed on little-endian systems. Fix
this for the new code.