Add quick check to see if own actor is fighting
Add a utility function to check if the player's own actor is in the actors list, and is currently fighting. This is
to be used as a quick check in sitations where the code will function correctly even if the result is no longer correct,
and where holding the actors list lock could block the list for longer than is desirable.