Clear actor pointers when releasing actors list
Add helper functions to set any actor pointers obtained by one of the lock_and_get_*() functions to NULL upon releasing
the actors list. Once the lock on the list is released, these pointers should no longer be used as they may be deleted
or otherwise modified. Setting them to NULL will help spot such uses after release.

The use of these functions is a convenience only, and is not enforced by the type system. An alternative, where
the lock_and_get_*() functions return a specialised wrapper type which has to be freed using a custom function,
was considered, but the use of the wrappers was too unwieldy, and the approach was not flexible enough. Users of
the actors list should therefore still excersise caution with actor pointers.